Split out the network-scripts
The network-scripts are now removed from the RHEL 8 base installation (and
possibly Fedora too). Let's remove them, so that we don't drag in the legacy
scripts when we don't need them (e.g. invoking through NetworkManager or
pptp and the like).

The boolean weak dependency is utilized to make life easier for those still
relying on the legacy network service and to avoid breakage on upgrades.