Backport patches from master for SSTP and EAP-TLS
With those changes will allow pppd to connect to an Azure VnetGWay using SSTP
(or using EAP-TLS to a Windows RAS server). Backport to Fedora endorsed
by Eivind Næss, their author and the author of sstp-client and
network-manager-sstp.

Links to git commits used to generate patches:
https://github.com/ppp-project/ppp/commit/e609ed8bb62e4648568eaa49fbbc858dfda6d122
https://github.com/ppp-project/ppp/commit/d706c95906d996534f13632a747af5dc617f306e
https://github.com/ppp-project/ppp/commit/d7e62a8499c4032d79e05afbd8fd3efd51c5b148
(the first one is only needed to make the second apply in the original
form from Git - those changed lines are removed anyway)